Fix: let error response in Repeater bubble up Determine if UnmarshallResponse represents a missing mandatory element Improve the diagnostic quality of some error messages

+++ src/main/java/info/rsdev/xb4j/model/converter/LongConverter.java
@@ -51,9 +51,9 @@ public class LongConverter implements IValueConverter {
 	}
 	
 	@Override
-	public Integer toObject(String value) {
+	public Long toObject(String value) {
 		if (value == null) { return null; }
-		return validator.isValid(Integer.valueOf(value));
+		return validator.isValid(Long.valueOf(value));
 	}
 	
 	@Override
